Plugin update

+++ var/plugins/hummingbird-performance/core/class-filesystem.php
@@ -153,19 +153,29 @@ class Filesystem {
 		// Removes CRITICAL Uncaught Error: Call to undefined function submit_button() in wp-admin/includes/file.php:1287.
 		require_once ABSPATH . 'wp-admin/includes/template.php';
 
-		// Check if the user has write permissions.
+		// Check if the user has 'write' permissions.
 		$access_type = get_filesystem_method();
 		if ( 'direct' === $access_type ) {
 			$this->fs_api = true;
 
 			// You can safely run request_filesystem_credentials() without any issues
 			// and don't need to worry about passing in a URL.
-			$credentials = request_filesystem_credentials( site_url() . '/wp-admin/', '', false, false, null );
+			$credentials = request_filesystem_credentials( site_url() . '/wp-admin/', '', false, false );
 
 			// Initialize the Filesystem API.
 			if ( ! WP_Filesystem( $credentials ) ) {
 				// Some problems, exit.
-				return new WP_Error( 'fs-error', __( 'Error: Unexpected error while writing a file. Please view error log for more information.', 'wphb' ) );
+				ob_start();
+				printf( /* translators: %1$s - code tag, %2$s - closing code tag, still having trouble link */
+					esc_html__( "Hummingbird has encountered an unexpected error while writing a file. To find out more, enable the WordPress debug log by adding the following line to your site’s wp-config.php file:%1\$sdefine('WP_DEBUG', true);%2\$s", 'wphb' ),
+					'<br><code>',
+					'</code><br><br>'
+				);
+				echo esc_html( '&nbsp;' );
+				Utils::still_having_trouble_link();
+				$text = ob_get_clean();
+
+				return new WP_Error( 'fs-error', $text );
 			}
 		} else {
 			// Don't have direct write access.
@@ -174,7 +184,14 @@ class Filesystem {
 
 		// Can not write to wp-content directory.
 		if ( defined( WP_CONTENT_DIR ) && ! is_writeable( WP_CONTENT_DIR ) ) {
-			return new WP_Error( 'fs-error', __( 'Error: The wp-content directory is not writable. Ensure the folder has proper read/write permissions for caching to function successfully.', 'wphb' ) );
+			return new WP_Error(
+				'fs-error',
+				sprintf( /* translators: %1$s - opening a tag, %2$s - closing a tag */
+					esc_html__( 'Your site’s wp-content directory is not writable. Please ensure the folder has the correct read and write %1$spermissions%2$s to ensure caching functions successfully.', 'wphb' ),
+					'<a href="https://wordpress.org/support/article/changing-file-permissions/#permission-scheme-for-wordpress" target="_blank">',
+					'</a>'
+				)
+			);
 		}
 
 		return true;
@@ -186,36 +203,104 @@ class Filesystem {
 	 * @since  1.7.2
 	 *
 	 * @access private
-	 * @param string $path  Path to delete.
+	 * @param string $path                 Path to delete.
+	 * @param bool   $skip_subdirectories  Skip subdirectories.
 	 *
 	 * @return bool
 	 */
-	private function native_dir_delete( $path ) {
+	private function native_dir_delete( $path, $skip_subdirectories = false ) {
 		if ( is_wp_error( $this->status ) ) {
 			return false;
 		}
 
 		// Use direct filesystem php functions.
-		$dir = @opendir( $path );
+		$dir = opendir( $path );
 
 		while ( false !== ( $file = readdir( $dir ) ) ) {
 			if ( '.' === $file || '..' === $file ) {
 				continue;
 			}
 
-			$full = $path . '/' . $file;
-			if ( is_dir( $full ) ) {
+			$full = trailingslashit( $path ) . $file;
+			if ( is_dir( $full ) && ! $skip_subdirectories ) {
 				$this->native_dir_delete( $full );
-			} else {
-				@unlink( $full );
+			} elseif ( ! is_dir( $full ) ) {
+				unlink( $full );
 			}
 		}
 
 		closedir( $dir );
-		@rmdir( $path );
+
+		// Remove empty directories or if allowed.
+		if ( 2 === count( scandir( $path ) ) || ! $skip_subdirectories ) {
+			rmdir( $path );
+		}
 
 		return true;
 	}
+
+	/**
+	 * Resolves minify module's asset path or url, taking into account user-defined path.
+	 *
+	 * @since 2.6.0
+	 *
+	 * @return string
+	 */
+	private function get_minify_asset_path() {
+		$upload = wp_upload_dir();
+		$path   = Settings::get_setting( 'file_path', 'minify' );
+
+		// Check if user defined a custom path.
+		if ( empty( $path ) ) {
+			return $upload['basedir'] . '/hummingbird-assets';
+		}
+
+		if ( strpos( $path, '/' ) === 0 ) { // root relative path.
+			return str_replace( '//', '/', ABSPATH . $path );
+		}
+
+		return trailingslashit( $upload['basedir'] ) . str_replace( './', '/', $path );
+	}
+
+	/**
+	 * Resolves directory info for critical assets like critical.css
+	 * Default path can be overridden by using global constants in wp-config.php.
+	 *
+	 * @since 2.6.0
+	 * @return array
+	 */
+	public static function critical_assets_dir() {
+		static $info;
+		if ( isset( $info ) ) {
+			return $info;
+		}
+
+		if ( defined( 'WPHB_CRITICAL_ASSETS_PATH' ) && ! empty( WPHB_CRITICAL_ASSETS_PATH ) ) {
+			$main_path = ABSPATH . trim( WPHB_CRITICAL_ASSETS_PATH, '\/ ' ) . '/';
+			if ( is_multisite() ) {
+				$blog_id     = get_current_blog_id();
+				$assets_path = $main_path . 'sites/' . $blog_id . '/';
+			} else {
+				$assets_path = $main_path;
+			}
+		} else {
+			$upload      = wp_upload_dir();
+			$assets_path = $upload['basedir'] . '/wphb-critical-assets/';
+		}
+
+		$rel_path = str_replace( ABSPATH, '', $assets_path );
+		if ( is_multisite() ) {
+			$assets_url = get_site_url( get_main_site_id(), $rel_path );
+		} else {
+			$assets_url = site_url( $rel_path );
+		}
+
+		return array(
+			'path' => $assets_path,
+			'url'  => trailingslashit( $assets_url ),
+		);
+	}
+
 
 	/**
 	 * Delete everything in selected folder.
@@ -223,13 +308,15 @@ class Filesystem {
 	 * @since  1.6.0
 	 * @since  1.7.2  Added if $this->fs_api check.
 	 * @since  1.9    Added $ao_module. If set to true will use the $dir path without $this->basedir
+	 * @since  2.7.3  Added $skip_subdirectories. @see https://incsub.atlassian.net/browse/HUM-497
+	 * @since  3.3.1  Removed $ao_module parameter in favor of purge_ao_cache()
 	 *
-	 * @param  string $dir        Directory in wp-content/wphb-cache/ to purge file from.
-	 * @param  bool   $ao_module  Asset Optimization module.
+	 * @param  string $dir                  Directory in wp-content/wphb-cache/ to purge file from.
+	 * @param  bool   $skip_subdirectories  Skip subdirectories.
 	 *
 	 * @return bool
 	 */
-	public function purge( $dir = 'cache', $ao_module = false ) {
+	public function purge( $dir = 'cache', $skip_subdirectories = false ) {
 		if ( is_wp_error( $this->status ) ) {
 			return false;
 		}
@@ -238,34 +325,7 @@ class Filesystem {
 			$dir = trailingslashit( $dir );
 		}
 
-		// Default behavior - use basedir path.
-		if ( ! $ao_module ) {
-			$path = $this->basedir . $dir;
-		} else {
-			$upload            = wp_upload_dir();
-			$user_defined_path = Settings::get_setting( 'file_path', 'minify' );
-			$basedir           = $upload['basedir'];
-
-			// Check if user defined a custom path.
-			if ( ! isset( $user_defined_path ) || empty( $user_defined_path ) ) {
-				$path = $basedir . '/hummingbird-assets';
-			} else {
-				if ( '/' === $user_defined_path[0] ) { // root relative path.
-					$custom_dir = ABSPATH . $user_defined_path;
-					$path       = str_replace( '//', '/', $custom_dir );
-				} else {
-					$user_defined_path = str_replace( './', '/', $user_defined_path );
-
-					// Prepend / to relative paths.
-					$prepend = '';
-					if ( '/' !== $user_defined_path[0] ) {
-						$prepend = '/';
-					}
-
-					$path = $upload['basedir'] . $prepend . $user_defined_path;
-				}
-			}
-		}
+		$path = $this->basedir . $dir;
 
 		// If directory not found - exit.
 		if ( ! is_dir( $path ) ) {
@@ -277,29 +337,75 @@ class Filesystem {
 			/**
 			 * WP_Filesystem global.
 			 *
-			 * @var WP_Filesystem_Base $wp_filesystem
+			 * @global WP_Filesystem_Base $wp_filesystem
 			 */
 			global $wp_filesystem;
 
 			// Delete all content inside the directory.
 			foreach ( $wp_filesystem->dirlist( $path ) as $asset ) {
+				// Skip subdirectories.
+				if ( $skip_subdirectories && isset( $asset['type'] ) && 'd' === $asset['type'] ) {
+					continue;
+				}
+
 				if ( ! $wp_filesystem->delete( $path . $asset['name'], true, $asset['type'] ) ) {
 					return false;
 				}
 			}
 
-			// Delete the directory itself.
-			if ( ! $wp_filesystem->delete( $path ) ) {
+			// Delete the directory itself if empty, or we can remove the dir.
+			if ( ( empty( $wp_filesystem->dirlist( $path ) ) || ! $skip_subdirectories ) && ! $wp_filesystem->delete( $path ) ) {
 				return false;
+			}
+
+			return true;
+		}
+
+		// Use direct filesystem php functions.
+		return $this->native_dir_delete( $path, $skip_subdirectories );
+	}
+
+	/**
+	 * Purge AO directories.
+	 *
+	 * @since 3.3.1  Abstracted from purge().
+	 *
+	 * @return void
+	 */
+	public function purge_ao_cache() {
+		if ( is_wp_error( $this->status ) ) {
+			return;
+		}
+
+		$path = trailingslashit( $this->get_minify_asset_path() );
+
+		// If directory not found - exit.
+		if ( ! is_dir( $path ) ) {
+			return;
+		}
+
+		// Use WP_Filesystem API to delete files.
+		if ( $this->fs_api ) {
+			/**
+			 * WP_Filesystem global.
+			 *
+			 * @global WP_Filesystem_Base $wp_filesystem
+			 */
+			global $wp_filesystem;
+
+			// Delete all content inside the directory.
+			foreach ( $wp_filesystem->dirlist( $path ) as $asset ) {
+				$wp_filesystem->delete( $path . $asset['name'], true, $asset['type'] );
+			}
+
+			// Delete the directory itself if empty.
+			if ( empty( $wp_filesystem->dirlist( $path ) ) ) {
+				$wp_filesystem->delete( $path );
 			}
 		} else {
 			// Use direct filesystem php functions.
-			if ( ! $this->native_dir_delete( $path ) ) {
-				return false;
-			}
+			$this->native_dir_delete( $path );
 		}
-
-		return true;
 	}
 
 	/**
@@ -320,7 +426,7 @@ class Filesystem {
 			/**
 			 * WP_Filesystem global.
 			 *
-			 * @var WP_Filesystem_Base $wp_filesystem
+			 * @global WP_Filesystem_Base $wp_filesystem
 			 */
 			global $wp_filesystem;
 
@@ -365,7 +471,7 @@ class Filesystem {
 			/**
 			 * WP_Filesystem global.
 			 *
-			 * @var WP_Filesystem_Base $wp_filesystem
+			 * @global WP_Filesystem_Base $wp_filesystem
 			 */
 			global $wp_filesystem;
 			return $wp_filesystem->exists( $path . $file );
@@ -414,13 +520,13 @@ class Filesystem {
 			/**
 			 * WP_Filesystem global.
 			 *
-			 * @var WP_Filesystem_Base $wp_filesystem
+			 * @global WP_Filesystem_Base $wp_filesystem
 			 */
 			global $wp_filesystem;
 
 			// Check if cache folder exists. If not - create it.
 			if ( ! $wp_filesystem->exists( $path ) ) {
-				if ( ! @wp_mkdir_p( $path ) ) {
+				if ( ! wp_mkdir_p( $path ) ) {
 					return new WP_Error(
 						'fs-dir-error',
 						sprintf(
@@ -447,7 +553,7 @@ class Filesystem {
 			// Use direct filesystem php functions.
 			// Check if cache folder exists. If not - create it.
 			if ( ! is_dir( $path ) ) {
-				if ( ! @wp_mkdir_p( $path ) ) {
+				if ( ! wp_mkdir_p( $path ) ) {
 					return new WP_Error(
 						'fs-dir-error',
 						sprintf(
@@ -518,7 +624,7 @@ class Filesystem {
 		$baseurl = $upload['baseurl'];
 
 		// Check if user defined a custom path.
-		if ( ! isset( $user_defined_path ) || empty( $user_defined_path ) ) {
+		if ( empty( $user_defined_path ) ) {
 			$custom_subdir = '/hummingbird-assets';
 			$custom_dir    = $upload['basedir'] . $custom_subdir;
 		} else {
@@ -554,7 +660,7 @@ class Filesystem {
 		 * TODO: For now, we're going to remove similar files. But better to just remove the wp_unique_filename function.
 		 */
 		if ( file_exists( trailingslashit( $custom_dir ) . $name ) ) {
-			@unlink( trailingslashit( $custom_dir ) . $name );
+			unlink( trailingslashit( $custom_dir ) . $name );
 		}
 
 		$filename = wp_unique_filename( $custom_dir, $name );
@@ -576,7 +682,7 @@ class Filesystem {
 			);
 		}
 
-		$ifp = @fopen( $new_file, 'wb' );
+		$ifp = fopen( $new_file, 'wb' );
 		if ( ! $ifp ) {
 			return array(
 				/* translators: %s: file name with path */
@@ -584,15 +690,15 @@ class Filesystem {
 			);
 		}
 
-		@fwrite( $ifp, $bits );
+		fwrite( $ifp, $bits );
 		fclose( $ifp );
 		clearstatcache();
 
 		// Set correct file permissions.
-		$stat  = @stat( dirname( $new_file ) );
+		$stat  = stat( dirname( $new_file ) );
 		$perms = $stat['mode'] & 0007777;
 		$perms = $perms & 0000666;
-		@chmod( $new_file, $perms );
+		chmod( $new_file, $perms );
 		clearstatcache();
 
 		// Compute the URL.
